before sleep

+++ commons/ios/smf_create_dmg_from_app/create_dmg.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #
 # Script to create a DMG from build App.
-# 
+#
 # The script creates a dmg from new app,
 # code signes the .dmg file
 #
@@ -21,29 +21,31 @@ export PATH
 
 CREATE_DMG=true
 CODE_SIGN_ID=
-TEMPLATE_PATH=
+USE_TEMPLATE=
+
 
 #
 # Variables
 #
 
 APPPATH=""
+TEMPLATE_PATH=""
 
 #
 # Helper
 #
 
 function usage() {
-	echo "Script to create a DMG from .app"
-	echo
-	echo "Required parameters:"
-	echo -e "--appPath, -p\t\t: Path to the .app"
-	echo
-	echo "Optional"
-	echo -e "--codesignid, -ci\t: Code signing identity, required if --codesign or -cs was used"
-	echo -e "--template, -t\t\t: Build the DMG from a template. If not specified, DMG will be created without a template"
-	echo
-	exit 0
+    echo "Script to create a DMG from .app"
+    echo
+    echo "Required parameters:"
+    echo -e "--appPath, -p\t\t: Path to the .app"
+    echo
+    echo "Optional"
+    echo -e "--codesignid, -ci\t: Code signing identity, required if --codesign or -cs was used"
+    echo -e "--template, -t\t\t: Build the DMG from a template. If not specified, DMG will be created without a template"
+    echo
+    exit 0
 }
 
 #
@@ -51,29 +53,34 @@ function usage() {
 #
 
 while [ $# -gt 0 ]; do
-	case "$1" in
-		--appPath | -p)
-			# Path where the App is stored after a successful build
-			APPPATH="$2"
-			shift 2
-			;;
-		--codesignid | -ci)
-			CODE_SIGN_ID=$2
-			shift 2
-			;;
-		--template | -t)
-			TEMPLATE_PATH=$2
-			shift
-			;;
-		-*)
-			usage
-			;;
-		*)
-	esac
+    case "$1" in
+        --appPath | -p)
+            # Path where the App is stored after a successful build
+            APPPATH="$2"
+            shift 2
+        ;;
+        --codesignid | -ci)
+            CODE_SIGN_ID=$2
+            shift 2
+        ;;
+        --template | -t)
+            TEMPLATE_PATH=$2
+            USE_TEMPLATE=true
+            shift
+        ;;
+        -*)
+            usage
+        ;;
+        *)
+    esac
 done
 
 if [ -z "$APPPATH" ]; then
-	usage
+    usage
+fi
+
+if [ "$USE_TEMPLATE" = "true" ] && [ -z "$TEMPLATE_PATH" ]; then
+    usage
 fi
 
 INFO_PLIST="$APPPATH"/Contents/Info.plist
@@ -84,61 +91,65 @@ SHORT_VERSION=$(defaults read "$INFO_PLIST" CFBundleVersion)
 APPDIR=$(dirname "$APPPATH")
 APPVERSION=${VERSION}-${SHORT_VERSION}
 APPFULLNAME=${NAME}-${APPVERSION}
-SRCFOLDER=${APPDIR}/${APPFULLNAME}
 
 #
 # DMG
 #
 
-# WITH TEMPLATE
-
-# Copy template
-
-# Convert template
-hdiutil convert -format UDSB -o testWritable.dmg test.dmg
-
-# Resize template
-hdiutil resize -size 200M testWritable.dmg.sparsebundle
-
-# Mount bundle - Returns path to volume
-
-$ORIGINAL_SPARSE_VOLUME_PATH=$(hdiutil attach testWritable.dmg.sparsebundle | egrep -o '/Volumes/[a-zA-Z]+$')
-
-# Rename Volume
-diskutil rename $ORIGINAL_SPARSE_VOLUME_PATH $VOLNAME
-
-# Empty dummy app
-rm -fr /Volumes/${VOLNAME}/HiDrive.app/*
-
-# Replace dummy app with real content - TODO: call the dummy app something generic
-ditto ${APPPATH} /Volumes/${VOLNAME}/HiDrive.app
-
-# Rename app
-mv /Volumes/${VOLNAME}/HiDrive.app /Volumes/${VOLNAME}/${NAME}.app
-
-# Detach
-hdiutil detach /Volumes/${VOLNAME}
-
-# Compact sparse bundle
-hdiutil compact testWritable.dmg.sparsebundle
-
-# Convert to final form
-hdiutil convert -format UDZO -o HiDrive_Alpha.dmg testWritable.dmg.sparsebundle
-
-# WITHOUT TEMPLATE
-
-if [ $CREATE_DMG = true ]; then
+if [ "$USE_TEMPLATE" = "true" ]; then
+    # WITH TEMPLATE
+	cd $APPDIR
+    
+    # Copy template
+    
+    # Convert template
+    hdiutil convert -format UDSB -o testWritable.dmg test.dmg
+    
+    # Resize template TODO: dynamic size ?
+    hdiutil resize -size 200M testWritable.dmg.sparsebundle
+    
+    # Mount bundle - Returns path to volume
+    
+    $ORIGINAL_SPARSE_VOLUME_PATH=$(hdiutil attach testWritable.dmg.sparsebundle | egrep -o '/Volumes/[a-zA-Z]+$')
+    
+    # Rename Volume
+    diskutil rename $ORIGINAL_SPARSE_VOLUME_PATH $VOLNAME
+    
+    # Empty dummy app content
+    rm -fr /Volumes/${VOLNAME}/HiDrive.app/*
+    
+    # Replace dummy app with real content - TODO: call the dummy app something generic
+    ditto ${APPPATH} /Volumes/${VOLNAME}/HiDrive.app
+    
+    # Rename app
+    mv /Volumes/${VOLNAME}/HiDrive.app /Volumes/${VOLNAME}/${NAME}.app
+    
+    # Detach
+    hdiutil detach /Volumes/${VOLNAME}
+    
+    # Compact sparse bundle
+    hdiutil compact testWritable.dmg.sparsebundle
+    
+    # Convert to final form
+    hdiutil convert -format UDZO -o HiDrive_Alpha.dmg testWritable.dmg.sparsebundle
+    
+    # Cleanup our trash TODO
+    
+else
+    # WITHOUT TEMPLATE
+	SRCFOLDER=${APPDIR}/${APPFULLNAME}
+    
     rm -rf "${SRCFOLDER}"
-	mkdir -p "${SRCFOLDER}"
-	cp -R "${APPPATH}" "${SRCFOLDER}"
-	cd "${SRCFOLDER}"
-	ln -s /Applications .
-	cd ..
-	hdiutil create "${NAME}".dmg -fs HFS+ -format UDZO -volname "${VOLNAME}" -srcfolder "${SRCFOLDER}"
-	if [ $? -gt 0 ]; then
-		echo "Abort: Error creating DMG"
-		exit 1
-	fi
+    mkdir -p "${SRCFOLDER}"
+    cp -R "${APPPATH}" "${SRCFOLDER}"
+    cd "${SRCFOLDER}"
+    ln -s /Applications .
+    cd ..
+    hdiutil create "${NAME}".dmg -fs HFS+ -format UDZO -volname "${VOLNAME}" -srcfolder "${SRCFOLDER}"
+    if [ $? -gt 0 ]; then
+        echo "Abort: Error creating DMG"
+        exit 1
+    fi
 fi
 
 #
@@ -146,11 +157,11 @@ fi
 #
 
 if [ -n "$CODE_SIGN_ID" ]; then
-	codesign -s "$CODE_SIGN_ID" "${APPDIR}"/"${NAME}".dmg
-	if [ $? -gt 0 ]; then
-		echo "Abort: Error code signing the DMG"
-		exit 1
-	fi
+    codesign -s "$CODE_SIGN_ID" "${APPDIR}"/"${NAME}".dmg
+    if [ $? -gt 0 ]; then
+        echo "Abort: Error code signing the DMG"
+        exit 1
+    fi
 fi
 
 # check if signed correctly
